Add filter/interlock to Ready To Launch mode
Filters out unwanted spuious interrupts from static that could to make the pattern skip a beat when the glass was rubbed.

Note that even thought we are now doing twice as many interrupts (rising and falling edges of `FOUT`) and a lot more tests, the run power of this mode is still only ~5.5uA indicating that interrupt overhead uses negliable energy, which is a little unexpected.