First working code generated by code!
Counts 0-9. Massive power savings - down to 4.4uA at 2Hz!

Still can optimize register loads and using idirect stores to squeeze that down a bit more.

Need to make it real with 2 digits for minutes and seconds, and also need to add lead in code to get us to a round hour.