Abort - too complicated
Ok, now we woudl need to figure out what the working_reg is and also check if the next step or any following step assumes it has that value, and then we need to load it in the setup so it will be ready if the optimimized code needs it. We are already at 95% code space usage, so not work saving a few instructions that only get run once after a battery change!