Increase coast time when changing batteries
Added a double diode to keep the RTC caps from back feeding into the MCU. This also offers reverse battery inesertion protection.

Connected FOE VBAT to prevent problems if this is floating when MCU is powered down.

Added  a resistor on the FOUT line to limit the current that backfeeds though the protect diode on the MCU GPIO pin when MCU is powered down.

TODO: This means that the firmware will need to leave the RTC MOS open.